gee test captcha code refactor

+++ authcaptcha/src/main/java/com/compose/authcaptcha/service/ApiService.kt
@@ -1,7 +1,5 @@
 package com.compose.authcaptcha.service
 
-import com.compose.authcaptcha.response.GeeTestResponse
-import okhttp3.Response
 import okhttp3.ResponseBody
 import retrofit2.http.GET
 import retrofit2.http.Query
@@ -10,6 +8,6 @@ interface ApiService {
 
     @GET("register-slide")
     suspend fun getCaptcha(
-        @Query("t") time: Int,
+        @Query("t") time: Long,
     ): ResponseBody
 }
